requirements.txt
autopep8==2.3.2
black==25.1.0
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
Flask==3.1.1
Flask-Mail==0.10.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mypy_extensions==1.1.0
mysql-connector-python==9.3.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pycodestyle==2.14.0
python-dotenv==1.1.1
requests==2.32.4
tomli==2.2.1
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
python-decouple
